html/interaction/focus/document-level-focus-apis/document-has-system-focus.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>HTML Test: focus - document has system focus</title>
<link rel="help" href="https://html.spec.whatwg.org/#has-focus-steps">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<input id="test">
<script>

promise_test(async t => {
  await new Promise(r => window.onload = r);
  assert_true(document.hasFocus(), "Document has focus on load.");

  const popup = window.open("support/popup.html", "otherwindow", "resizable");
  let gotBlur = false;
  window.onblur = () => gotBlur = true;
  const msg = await new Promise(r => window.onmessage = ({data}) => r(data));
  assert_equals(msg, "visible",
                "Test requires popups be opened (may require harness flags)");
  assert_true(gotBlur, "Document received blur event when popup opened");
  assert_false(document.hasFocus(), "Document lost focus when popup opened");

  popup.close();
  await new Promise(r => window.onfocus = r);
  assert_true(true, "Document received focus event when popup closed");
  assert_true(document.hasFocus(), "Document regained focus when popup closed");
}, "Top-level document receives blur/focus events and loses system focus " +
   "during opening/closing of a popup");

</script>
